refactor: Organizando services
