+++ src/app/guard/auth.guard.ts
@@ -1,6 +1,6 @@
 import { CanActivateFn, Router } from '@angular/router';
 import { inject } from '@angular/core';
-import { AlertService } from '../services/alert.service';
+import { AlertService } from '../services/componsents/alert.service';
 import { AuthService } from '../services/auth.service';
 
 export const authGuard: CanActivateFn = (route, state) => {
