+++ src/app/shared/components/global-alerts/global-alerts.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, TemplateRef, ViewChild, inject } from '@angular/core';
-import { AlertService } from '../../../services/alert.service';
+import { AlertService } from '../../../services/componsents/alert.service';
 import { NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';
 import { AlertsComponent } from '../alerts/alerts.component';
